[FSDP] summon_full_params fix
Needs to synchronize all_gather stream. Added test fails without this
fix

Differential Revision: [D34430602](https://our.internmc.facebook.com/intern/diff/D34430602/)

[ghstack-poisoned]
